database

+++ app/src/main/java/com/loan555/myviewpager/model/DataNoteItem.kt
@@ -3,7 +3,7 @@ package com.loan555.myviewpager.model
 import java.io.Serializable
 
 class DataNoteItem(
-    var noteId: Int,
+    var noteId: Long,
     var date: CalendarDateModel,
     var titleHead: String,
     var titleBody: String
